+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.2.0.4 (upcoming)
 
 * Added Tenant Name variable being able to have Spark Dispatcher with different name
+* Enable Debug mode
 
 ## 2.2.0.3 (December 27, 2017)